Move Send/Receive/Close/CloseOutputAsync argument validation into PAL
We want the arg validation shared across uses of the managed web socket across repos.
