+++ src/System.Net.WebSockets.Client/src/System/Net/WebSockets/WebSocketHandle.Windows.cs
@@ -66,6 +66,20 @@ namespace System.Net.WebSockets
             bool endOfMessage,
             CancellationToken cancellationToken)
         {
+            if (messageType != WebSocketMessageType.Text && messageType != WebSocketMessageType.Binary)
+            {
+                string errorMessage = SR.Format(
+                        SR.net_WebSockets_Argument_InvalidMessageType,
+                        nameof(WebSocketMessageType.Close),
+                        nameof(SendAsync),
+                        nameof(WebSocketMessageType.Binary),
+                        nameof(WebSocketMessageType.Text),
+                        nameof(CloseOutputAsync));
+                throw new ArgumentException(errorMessage, nameof(messageType));
+            }
+
+            WebSocketValidate.ValidateArraySegment(buffer, nameof(buffer));
+
             return _webSocket.SendAsync(buffer, messageType, endOfMessage, cancellationToken);
         }
 
@@ -73,6 +87,7 @@ namespace System.Net.WebSockets
             ArraySegment<byte> buffer,
             CancellationToken cancellationToken)
         {
+            WebSocketValidate.ValidateArraySegment(buffer, nameof(buffer));
             return _webSocket.ReceiveAsync(buffer, cancellationToken);
         }
 
@@ -81,6 +96,7 @@ namespace System.Net.WebSockets
             string statusDescription,
             CancellationToken cancellationToken)
         {
+            WebSocketValidate.ValidateCloseStatus(closeStatus, statusDescription);
             return _webSocket.CloseAsync(closeStatus, statusDescription, cancellationToken);
         }
 
@@ -89,6 +105,7 @@ namespace System.Net.WebSockets
             string statusDescription,
             CancellationToken cancellationToken)
         {
+            WebSocketValidate.ValidateCloseStatus(closeStatus, statusDescription);
             return _webSocket.CloseOutputAsync(closeStatus, statusDescription, cancellationToken);
         }
